artifacts_managment/pyproject.toml
[project]
name = "artifacts-managment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fpdf>=1.7.2",
    "google-adk>=1.25.1",
    "pandas>=3.0.1",
    "pypandoc>=1.16.2",
    "rich>=14.3.3",
    "tabulate>=0.9.0",
]
